+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 #Changelog
 
+## 1.xx-dev
+Release date: ??.??.????
+
+
+
 ## 1.26
 Release date: 01.11.2018