+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Release date: ??.??.????
 
 * Support for new guide.teacher.helsinki.fi domain (HUB-381)
 * Added a user group: Teachers (HUB-371)
+* Breadcrumb (HUB-388)
 
 
 ## 1.26