+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Release date: ??.??.????
 * Support for new guide.teacher.helsinki.fi domain (HUB-381)
 * Added a user group: Teachers (HUB-371)
 * Breadcrumb (HUB-388)
+* Removed top content (HUB-384)
 
 
 ## 1.26