+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Release date: ??.??.????
 * Support for new guide.teacher.helsinki.fi domain (HUB-381)
 * Added a user group: Teachers (HUB-371)
 * Breadcrumb (HUB-388)
+* Support for referring to student articles from the teacher domain (HUB-374)
 
 
 ## 1.26