+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Release date: ??.??.????
 * Breadcrumb (HUB-388)
 * Removed top content (HUB-384)
 * Support for referring to student articles from the teacher domain (HUB-374)
+* Display only the login/logout main menu item on teacher domain (HUB-397)
 
 
 ## 1.26