+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Release date: ??.??.????
 * Breadcrumb (HUB-388)
 * Removed top content (HUB-384)
 * Support for referring to student articles from the teacher domain (HUB-374)
+* Copy URL button position change (HUB-394)
+* Display only the login/logout main menu item on teacher domain (HUB-397)
 * Display general office hours grouped by language (HUB-362).