+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Release date: ??.??.????
 * Copy URL button position change (HUB-394)
 * Display only the login/logout main menu item on teacher domain (HUB-397)
 * Display general office hours grouped by language (HUB-362).
+* Do not display news blocks on teacher domain frontpage (HUB-392)
 
 
 ## 1.26