+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Release date: ??.??.????
 * Display only the login/logout main menu item on teacher domain (HUB-397)
 * Display general office hours grouped by language (HUB-362).
 * Do not display news blocks on teacher domain frontpage (HUB-392)
+* IE11: Fixed overflowing news and theme titles (HUB-400, HUB-404)
+* Support for referring to student articles and themes from the teacher domain (HUB-373)
 * News archive: Condensed degree programme list (HUB-405)