+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Release date: ??.??.????
 * IE11: Fixed overflowing news and theme titles (HUB-400, HUB-404)
 * Support for referring to student articles and themes from the teacher domain (HUB-373)
 * News archive: Condensed degree programme list (HUB-405)
+* Link to student news (HUB-376)
 
 
 ## 1.26