+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Release date: ??.??.????
 * Do not display news blocks on teacher domain frontpage (HUB-392)
 * IE11: Fixed overflowing news and theme titles (HUB-400, HUB-404)
 * Support for referring to student articles and themes from the teacher domain (HUB-373)
+* News archive: Condensed degree programme list (HUB-405)
+* Link to student news (HUB-376)
+* Front page: Teaching themes listing (HUB-383)
 * Domain-specific search results (HUB-375)