+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Release date: ??.??.????
 * Link to student news (HUB-376)
 * Front page: Teaching themes listing (HUB-383)
 * Domain-specific search results (HUB-375)
+* Domain-specific front page title and text blocks (HUB-382)
 
 
 ## 1.26